[NUI][Badges] Adding comments and README file
