+++ Mobile/NUI/Badges/README.md
@@ -1,0 +1,18 @@
+# Badges
+Badges is a sample application which demonstrates how to manage applications badge
+counter using [Tizen Applications API](https://samsung.github.io/TizenFX/stable/api/Tizen.Applications.html).
+
+![Main page](./Screenshots/main.png)
+
+### Features
+* List of changeable applications (with current badge value)
+* List of other applications (with current badge value)
+* Changing badge counter value
+
+### Prerequisites
+
+* [Visual Studio](https://www.visualstudio.com/) - Buildtool, IDE
+* [Visual Studio Tools for Tizen](https://docs.tizen.org/application/vstools/install) - Visual Studio plugin for Tizen .NET application development
+
+### Author
+* Bartłomiej Uliasz
